dockerfiles: add missing host tools to base
bitbake now checks to see if we have a certain set of host tools
available for before it will launch a build. This check caught several
missing tools that are mostly used for testimage.  This commit adds
the missing tools.

What was missing:
opensuse-<all>  - hostname , ip
ubuntu-16-<all> - ping

Signed-off-by: brian avery <brian.avery@intel.com>