smoke.sh: run smoke tests container as user uid
Some of the smoke tests put files in the $workdir which is set to
/tmp/tmp.GUID_smoke. This workdir is mounted from the host into the
container. If the tests are successful, this directory's contents are
removed. Previously, these tests were running as the default user 1000
in the container so if you were uid=1010 on the system, the smoke tests
would drop files in /tmp/tmp.GUID_smoke owned by uid=1000 and the final
smoke test cleanup to remove all the files from the /tmp/tmp.GUID_smoke
would fail.  Similarly, it's better to run as the user uid in case they
want to edit/modify/whatever the files.

Signed-off-by: brian avery <brian.avery@intel.com>