all-builder: switch to templates for -builder dirs
This commit removes the Dockerfiles from the -builder directories and
relies on build_containers.sh populating the Dockerfiles from the
dockerfiles/templates/Dockerfile.builder.  All the files are the same
except for which -base distro they pull from.

A README is also added to the -builder directories to discourage people
from putting Dockerfiles there as the current logic will just overwrite them.

Signed-off-by: brian avery <brian.avery@intel.com>