build_container.sh: Don't use the builder directories
This changes build_container.sh to longer depend on the existence of a
"builder" directory. Previously, the template dockerfile was being
copied into the builder directory. However, this isn't necessary because
a temporary directory is created to be used for the build.

So now just copy the template to the temporary directory, and do the sed
replacement there. This allows for the removal of a lot of useless
directories and files.

Signed-off-by: Randy Witt <randy.e.witt@intel.com>