centos-7: Fix vnc test failure
Patch /usr/bin/vncserver to no longer run "vncserver -kill $DISPLAY".

The vnc smoke test was failing for Centos 7. For some reason the default
xstartup file created by /usr/bin/vncserver would run "vncserver -kill
$DISPAY".

This command is not in the upstream tigervnc and it was having the
effect of killing the vncserver right after it was started.

Signed-off-by: Randy Witt <randy.e.witt@intel.com>