fedora-32: Fix vnc test failures
The latest version of tigervnc no longer supports running using
vncserver. This caused the tests in vnc-in-container-test.sh to fail
since vncserver was being used.

Since vnc test is just testing that a vnc server instance can be started
and used, instead use Xvnc directly. This ensure the test can be run
successfully on all the distros.

Signed-off-by: Randy Witt <randy.e.witt@intel.com>