fedora: Add Fedora 40
Based on the Fedora 39 Dockerfile.

Fedora 40 replaces wget with the re-write wget2.  This drop-in
replacement does not yet have support for FTP. As a result the
bitbake fetcher fails when it uses the wget provided by Fedora
with --passive-ftp.

The workaround is to use the poky build tools to use the
wget-native provided by the Yocto Project.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>